TB2016 - masking minor changes
git-svn-id: https://belle2.cc.kek.jp/svn/trunk/software@27784 b38c6c8f-bee8-465d-8448-9abe6915d496

+++ testbeam/vxd/modules/src/MaskingModule.cc
@@ -443,7 +443,7 @@ void MaskingModule::event()
 
   StoreObjPtr<EventMetaData> storeEventMetaData;
 
-  if (storeEventMetaData->getEvent() > m_nEventsProcess) {
+  if (storeEventMetaData->getEvent() > (unsigned int)m_nEventsProcess) {
     storeEventMetaData->setEndOfData();
   }
   m_StatAllEvents++;
